升星数据修改； 升星点击continue之后才发请求；
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@646545 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/star_up_done_entity.dart
@@ -112,14 +112,14 @@ class StarUpDoneTeamPlayerVOMaxAbility {
 
 @JsonSerializable()
 class StarUpDoneTeamPlayerVOPotential {
-	late int blk = 0;
+	late double blk = 0;
 	late double tech;
-	late int pass = 0;
-	late int stl = 0;
-	late int threePm = 0;
-	late int ftm = 0;
-	late int pts = 0;
-	late int reb = 0;
+	late double pass = 0;
+	late double stl = 0;
+	late double threePm = 0;
+	late double ftm = 0;
+	late double pts = 0;
+	late double reb = 0;
 
 	StarUpDoneTeamPlayerVOPotential();
